chat-insights/requirements.txt
streamlit
langchain
faiss-cpu
openai
python-dotenv
langchain-community
langchain-openai
tiktoken
